Add mute state change event and navigation view to MainWindow
- Introduced an event for mute state changes in App.xaml.cs to facilitate communication between components.
- Updated MainWindow.xaml to replace the previous layout with a NavigationView for improved navigation and user experience.
- Adjusted MainWindow.xaml.cs to handle navigation and mute state updates, enhancing the application's responsiveness and organization.
- Removed obsolete code related to the previous UI structure, streamlining the codebase.
