+++ Pages/AboutPage.xaml
@@ -1,0 +1,52 @@
+<?xml version="1.0" encoding="utf-8"?>
+<Page
+    x:Class="silence_.Pages.AboutPage"
+    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+    xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+    xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+    mc:Ignorable="d"
+    Background="Transparent">
+
+    <ScrollViewer VerticalScrollBarVisibility="Auto" 
+                  HorizontalScrollBarVisibility="Disabled">
+        <StackPanel Spacing="24" Margin="24,24,24,24">
+            <StackPanel Spacing="8" HorizontalAlignment="Center">
+                <Image Source="/Assets/app.png" 
+                       Width="64" 
+                       Height="64"
+                       HorizontalAlignment="Center"/>
+                <TextBlock Text="silence!" 
+                           FontSize="28" 
+                           FontWeight="Bold"
+                           HorizontalAlignment="Center"/>
+                <TextBlock Text="v1.0.0" 
+                           FontSize="14"
+                           Opacity="0.6"
+                           HorizontalAlignment="Center"/>
+            </StackPanel>
+            
+            <TextBlock Text="A simple microphone mute/unmute utility with global hotkey support."
+                       TextWrapping="Wrap"
+                       TextAlignment="Center"
+                       Opacity="0.8"/>
+            
+            <StackPanel Spacing="8" HorizontalAlignment="Center">
+                <HyperlinkButton Content="View on GitHub"
+                                 NavigateUri="https://github.com/vertopolkaLF/silence"
+                                 HorizontalAlignment="Center"/>
+                <TextBlock Text="Made with ❤️ by vertopolkaLF" 
+                           FontSize="12"
+                           Opacity="0.5"
+                           HorizontalAlignment="Center"/>
+            </StackPanel>
+            
+            <StackPanel Spacing="4" Margin="0,20,0,0">
+                <TextBlock Text="Credits" FontWeight="SemiBold" FontSize="14"/>
+                <TextBlock Text="• WinUI 3 - Microsoft" Opacity="0.7" FontSize="12"/>
+                <TextBlock Text="• H.NotifyIcon - havendv" Opacity="0.7" FontSize="12"/>
+            </StackPanel>
+        </StackPanel>
+    </ScrollViewer>
+</Page>
+
