+++ Pages/AboutPage.xaml.cs
@@ -1,0 +1,13 @@
+using Microsoft.UI.Xaml.Controls;
+
+namespace silence_.Pages
+{
+    public sealed partial class AboutPage : Page
+    {
+        public AboutPage()
+        {
+            InitializeComponent();
+        }
+    }
+}
+
