+++ Pages/GeneralPage.xaml
@@ -1,0 +1,132 @@
+<?xml version="1.0" encoding="utf-8"?>
+<Page
+    x:Class="silence_.Pages.GeneralPage"
+    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+    xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+    xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+    mc:Ignorable="d"
+    Background="Transparent">
+
+    <ScrollViewer VerticalScrollBarVisibility="Auto" 
+                  HorizontalScrollBarVisibility="Disabled">
+        <StackPanel x:Name="ContentPanel" Spacing="20" Margin="24,24,24,24">
+        
+            <!-- Compact header with mute status -->
+            <Grid Height="50">
+                <Grid.ColumnDefinitions>
+                    <ColumnDefinition Width="Auto"/>
+                    <ColumnDefinition Width="*"/>
+                </Grid.ColumnDefinitions>
+            
+                <Border x:Name="MuteButton" 
+                        Grid.Column="0"
+                        Tapped="MuteButton_Click"
+                        PointerEntered="MuteButton_PointerEntered"
+                        PointerExited="MuteButton_PointerExited"
+                        PointerPressed="MuteButton_PointerPressed"
+                        PointerReleased="MuteButton_PointerReleased"
+                        Width="50" Height="50"
+                        CornerRadius="25"
+                        VerticalAlignment="Center"
+                        Background="#28A745">
+                    <FontIcon x:Name="MuteIcon" 
+                              FontFamily="{ThemeResource SymbolThemeFontFamily}" 
+                              Glyph="&#xE720;" 
+                              FontSize="24"
+                              Foreground="White"
+                              HorizontalAlignment="Center"
+                              VerticalAlignment="Center"/>
+                </Border>
+            
+                <StackPanel Grid.Column="1" 
+                            Orientation="Horizontal" 
+                            VerticalAlignment="Center"
+                            Margin="16,0,0,0"
+                            Spacing="4">
+                    <TextBlock Text="Microphone is" 
+                               FontSize="18" 
+                               FontWeight="SemiBold"
+                               VerticalAlignment="Center"/>
+                    <!-- Animated text container with clipping -->
+                    <Grid Height="26" VerticalAlignment="Center">
+                        <Grid.Clip>
+                            <RectangleGeometry Rect="0,0,200,26"/>
+                        </Grid.Clip>
+                        <TextBlock x:Name="MuteStatusText" 
+                                   Text="unmuted" 
+                                   FontSize="18" 
+                                   FontWeight="SemiBold"
+                                   VerticalAlignment="Center"
+                                   HorizontalAlignment="Left">
+                            <TextBlock.RenderTransform>
+                                <CompositeTransform x:Name="MuteStatusTransform"/>
+                            </TextBlock.RenderTransform>
+                        </TextBlock>
+                        <TextBlock x:Name="MuteStatusTextAlt" 
+                                   Text="muted" 
+                                   FontSize="18" 
+                                   FontWeight="SemiBold"
+                                   VerticalAlignment="Center"
+                                   HorizontalAlignment="Left"
+                                   Opacity="0">
+                            <TextBlock.RenderTransform>
+                                <CompositeTransform x:Name="MuteStatusTransformAlt" TranslateY="30"/>
+                            </TextBlock.RenderTransform>
+                        </TextBlock>
+                    </Grid>
+                </StackPanel>
+            </Grid>
+
+            <!-- Microphone Selection -->
+            <StackPanel Spacing="8">
+                <TextBlock Text="Microphone" FontWeight="SemiBold" FontSize="14"/>
+                <ComboBox x:Name="MicrophoneComboBox"
+                          HorizontalAlignment="Stretch"
+                          SelectionChanged="MicrophoneComboBox_SelectionChanged"
+                          PlaceholderText="Select microphone..."/>
+            </StackPanel>
+
+            <!-- Hotkey Configuration -->
+            <StackPanel Spacing="8">
+                <TextBlock Text="Hotkey" FontWeight="SemiBold" FontSize="14"/>
+                <Grid ColumnSpacing="12">
+                    <Grid.ColumnDefinitions>
+                        <ColumnDefinition Width="*"/>
+                        <ColumnDefinition Width="Auto"/>
+                    </Grid.ColumnDefinitions>
+                    <TextBox x:Name="HotkeyTextBox" 
+                             Grid.Column="0"
+                             IsReadOnly="True"
+                             PlaceholderText="Press a key..."
+                             Text="Ctrl + Alt + M"/>
+                    <Button x:Name="RecordHotkeyButton" 
+                            Grid.Column="1"
+                            Content="Record"
+                            Click="RecordHotkeyButton_Click"/>
+                </Grid>
+                <CheckBox x:Name="IgnoreModifiersCheckBox"
+                          Content="Ignore modifiers (Shift, Ctrl, Alt, Win)"
+                          IsChecked="False"
+                          Checked="IgnoreModifiersCheckBox_Changed"
+                          Unchecked="IgnoreModifiersCheckBox_Changed"/>
+            </StackPanel>
+
+            <!-- Startup Options -->
+            <StackPanel Spacing="8">
+                <TextBlock Text="Startup" FontWeight="SemiBold" FontSize="14"/>
+                <CheckBox x:Name="AutoStartCheckBox"
+                          Content="Start with Windows"
+                          Checked="AutoStartCheckBox_Changed"
+                          Unchecked="AutoStartCheckBox_Changed"/>
+                <CheckBox x:Name="StartMinimizedCheckBox"
+                          Content="Start minimized to tray"
+                          IsChecked="False"
+                          Checked="StartMinimizedCheckBox_Changed"
+                          Unchecked="StartMinimizedCheckBox_Changed"/>
+            </StackPanel>
+        
+        </StackPanel>
+    </ScrollViewer>
+</Page>
+
